Internal: Rename scripts/util to scripts/utils.
