+++ scripts/utils/assertisinstalled.js
